Removed sprite sheets.
The p5.play API works with sprite sheets via the loadSpriteSheet()
method. For right now it's too cumbersome to use because you have to
figure out the location (x,y) and dimensions of the sprite within the
sheet. This involves manually inspecting the sheet in something like
GIMP.

The advantage of sprite sheets is that you won't have to deploy so many
image assets for png sequences.