Stop relying on p5 typings supplied by the p5.vscode extension.
This change generalizes the repo to be an npm project. This allows us to
manage dependences in package.json and remove any dependency on local
user directories for typings. For projects created with p5.vscode, the
resulting jsconfig.json includes typings that are stored in a directory
relative to the user-level extension for example:

  `$USER/.vscode/extension/samplavigne.p5-vscode-1.2.4/p5types/`

The implication is that a project created with the extension cannot be
cloned without updating this line, making it hard to create and
distribute projects for teaching purposes.

After this change, p5 typings dependency will be managed in
package.json. The project will look for typings in project-level
node_modules dir.

So to use the project you need to complete these steps:

1. Clone the repo
2. Then run `npm install`